+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.12] - 2026-04-09
+### Changed
+- WIP: backup distro
+- WIP: List newer files in TA
+
 ## [1.0.11] - 2026-04-08
 ### Changed
 - WIP: Backup TA (last checked date see RTM)